NextID is now just a SubCommand instead of a full Target
much simpler!
