+++ SiteApp.xcodeproj/xcshareddata/xcschemes/site_tool.xcscheme
@@ -61,6 +61,10 @@
             isEnabled = "NO">
          </CommandLineArgument>
          <CommandLineArgument
+            argument = "nextID"
+            isEnabled = "NO">
+         </CommandLineArgument>
+         <CommandLineArgument
             argument = "https://www.bolsinga.com/json/"
             isEnabled = "YES">
          </CommandLineArgument>
