+++ Sources/site_tool/NextIDCommand.swift
@@ -1,0 +1,42 @@
+//
+//  NextIDCommand.swift
+//
+//  Created by Greg Bolsinga on 8/3/23.
+//
+
+import ArgumentParser
+import Foundation
+
+extension Vault {
+  fileprivate var nextShowID: Show.ID {
+    let nextIndex = max(concertMap.count, 0)
+    return "\(ArchivePath.showPrefix)\(nextIndex)"
+  }
+
+  fileprivate var nextVenueID: Venue.ID {
+    let nextIndex = max(venueDigestMap.count, 0)
+    return "\(ArchivePath.venuePrefix)\(nextIndex)"
+  }
+
+  fileprivate var nextArtistID: Venue.ID {
+    let nextIndex = max(artistDigestMap.count, 0)
+    return "\(ArchivePath.artistPrefix)\(nextIndex)"
+  }
+}
+
+struct NextIDCommand: AsyncParsableCommand {
+  static let configuration = CommandConfiguration(
+    commandName: "nextID",
+    abstract: "Prints the next ID for all the music entires."
+  )
+
+  @OptionGroup var rootURL: RootURLArguments
+
+  func run() async throws {
+    let vault = try await rootURL.vault()
+
+    print("Next Show: \(vault.nextShowID)")
+    print("Next Venue: \(vault.nextVenueID)")
+    print("Next Artist: \(vault.nextArtistID)")
+  }
+}
