+++ Sources/site_tool/Program.swift
@@ -14,7 +14,7 @@ struct Program: AsyncParsableCommand {
     commandName: "site_tool",
     abstract: "A tool for working with site data.",
     subcommands: [
-      DumpVaultCommand.self, DumpDiaryCommand.self, DumpBracketCommand.self,
+      DumpVaultCommand.self, DumpDiaryCommand.self, DumpBracketCommand.self, NextIDCommand.self
     ],
     defaultSubcommand: DumpVaultCommand.self
   )
